src/requirements.txt
# Core Framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# PyTorch (CPU version)
--index-url https://download.pytorch.org/whl/cpu
torch>=2.3.0
torchaudio>=2.3.0

# RAG stack
sentence-transformers==2.7.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-huggingface>=0.0.1
faiss-cpu>=1.8.0
transformers>=4.41.0
accelerate>=0.30.0

# Document loaders
pypdf>=4.2.0
docx2txt>=0.8

# F5-TTS
f5-tts @ git+https://github.com/SWivid/F5-TTS.git

# Audio & Media
soundfile>=0.12.1
numpy>=1.26.0

# Document Generation với hình ảnh 
python-docx>=1.1.0
Pillow>=10.0.0
reportlab>=4.0.0
requests>=2.31.0